Refactored to split Geocoding functionality off into it's own class
Separation of concerns and all that, init.
