+++ DmgLocationUtils/src/main/java/com/dmgdesignuk/locationutils/easylocationutility/DeviceLocationCallback.java
@@ -1,4 +1,4 @@
-package com.dmgdesignuk.devicelocationutility;
+package com.dmgdesignuk.locationutils.easylocationutility;
 
 import android.location.Location;
