+++ DmgLocationUtils/src/main/java/com/dmgdesignuk/locationutils/easylocationutility/EasyLocationUtility.java
@@ -1,4 +1,4 @@
-package com.dmgdesignuk.devicelocationutility;
+package com.dmgdesignuk.locationutils.easylocationutility;
 
 import android.Manifest;
 import android.annotation.SuppressLint;
@@ -8,13 +8,13 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.IntentSender;
 import android.content.pm.PackageManager;
-import android.location.Address;
-import android.location.Geocoder;
 import android.location.Location;
 import android.support.annotation.NonNull;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.FragmentActivity;
 import android.util.Log;
+
+import com.dmgdesignuk.locationutils.R;
 
 import com.google.android.gms.common.api.ResolvableApiException;
 import com.google.android.gms.location.FusedLocationProviderClient;
@@ -29,39 +29,30 @@ import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 
-import java.io.IOException;
 import java.lang.ref.WeakReference;
-import java.util.List;
-import java.util.Locale;
 
 
 /**
  * <h1>Wraps interactions with the Google Play Services FusedLocationProviderClient</h1>
- * <p>
- *  Provides an easy way to work with a device's location, including testing for and
- *  requesting the necessary permission(s) as needed.
- * </p>
- * <p>
- *  Provides several methods for retrieving the device's current or last known location.
- *  Also provides methods for obtaining the address details of a given location via reverse
- *  Geocoding.
- * </p>
- * <p>
- *  Requires permission to access device's fine location. Make sure to add the appropriate permission
- *  request in your app's Manifest.
- * </P>
- * <p>
- *  The calling Activity should implement <strong>ActivityCompat.OnRequestPermissionsResultCallback</strong>
- *  to receive and handle the result(s) of a permission request.
- * </P>
+ *
+ * <p>Provides an easy way to work with a device's location, including testing for and
+ *    requesting the necessary permission(s) as needed.</p>
+ *
+ * <p>Provides several methods for retrieving the device's current or last known location.</p>
+ *
+ * <p>Requires permission to access device's fine location. Make sure to add the appropriate permission
+ *    request in your app's Manifest.</P>
+ *
+ * <p>The calling Activity should implement <strong>ActivityCompat.OnRequestPermissionsResultCallback</strong>
+ *    to receive and handle the result(s) of a permission request.</P>
  *
  * @author  Dave Gibbons (dave@dmgdesignuk.com)
  * @version 0.1.0
  * @since	2018-08-14
  */
-public class DeviceLocationUtility extends LocationCallback
+public class EasyLocationUtility extends LocationCallback
 {
-    private static final String TAG = DeviceLocationUtility.class.getSimpleName();
+    private static final String TAG = EasyLocationUtility.class.getSimpleName();
 
 
     /**
@@ -75,24 +66,6 @@ public class DeviceLocationUtility extends LocationCallback
         public static final int SMART_LOCATION = 3;
     }
 
-    /**
-     * Static nested class defining int constants for use with reverse Geocoding of address data
-     */
-    public static class AddressCodes
-    {
-        public static final int ADMIN_AREA = 0;
-        public static final int CITY_NAME = 1;
-        public static final int COUNTRY_CODE = 2;
-        public static final int COUNTRY_NAME = 3;
-        public static final int FEATURE_NAME = 4;
-        public static final int FULL_ADDRESS = 5;
-        public static final int PHONE_NUMBER = 6;
-        public static final int POST_CODE = 7;
-        public static final int PREMISES = 8;
-        public static final int STREET_ADDRESS = 9;
-        public static final int SUB_ADMIN_AREA = 10;
-        public static final int SUB_THOROUGHFARE = 11;
-    }
 
     /**
      * Static nested class to handle displaying a dialog providing additional rationale to
@@ -184,7 +157,6 @@ public class DeviceLocationUtility extends LocationCallback
     private Context mContext;
     private LocationRequest mLocationRequest;
     private LocationCallback mLocationCallback;
-    private Geocoder mGeocoder;
 
     // Flag to track if continuous location updates have been requested at any point within the
     // object's life. False by default and set to true by any method that requests continuous updates.
@@ -205,7 +177,7 @@ public class DeviceLocationUtility extends LocationCallback
      *
      * Default LocationRequest parameters are also assigned here.
      */
-    public DeviceLocationUtility(Activity activity){
+    public EasyLocationUtility(Activity activity){
         // assign the activity to the weak reference
         this.weakActivity = new WeakReference<>(activity);
 
@@ -329,7 +301,7 @@ public class DeviceLocationUtility extends LocationCallback
      * @param   callback An interface which must be implemented by the caller in order to
      *		    receive the results of the location request.
      *
-     * @see DeviceLocationUtility#stopLocationUpdates()
+     * @see EasyLocationUtility#stopLocationUpdates()
      */
     @SuppressLint("MissingPermission")
     public void getCurrentLocationUpdates(final DeviceLocationCallback callback){
@@ -509,7 +481,7 @@ public class DeviceLocationUtility extends LocationCallback
      *
      * @param   requestCode the request code passed in by requestPermission()
      *
-     * @see DeviceLocationUtility#requestPermission(int)
+     * @see EasyLocationUtility#requestPermission(int)
      */
     private void startPermissionRequest(int requestCode){
         // Re-acquire a strong reference to the calling activity and verify that it still exists and is active
@@ -634,7 +606,7 @@ public class DeviceLocationUtility extends LocationCallback
      * This should be called in the calling Activity's onResume() method if
      * you want your app to continue to receive location updates when it resumes
      *
-     * @see DeviceLocationUtility#stopLocationUpdates()
+     * @see EasyLocationUtility#stopLocationUpdates()
      */
     @SuppressLint("MissingPermission")
     public void resumeLocationUpdates(){
@@ -672,132 +644,4 @@ public class DeviceLocationUtility extends LocationCallback
     }
 
 
-    /**
-     * Returns a List object containing the address information for the supplied
-     * Location object. Will be null if no address found. The caller should
-     * implement an AddressResultCallback to receive and handle the result.
-     *
-     * @param   location    A Location object containing latitude and longitude.
-     * @param   callback    An interface which must be implemented by the caller in order to
-     *		                receive the results of the request.
-     */
-    public void getAddressList(Location location, AddressResultCallback callback){
-
-        mGeocoder = new Geocoder(mContext, Locale.getDefault());
-        List<Address> addressList = null;
-
-        try {
-
-            addressList = mGeocoder.getFromLocation(
-                    location.getLatitude(), location.getLongitude(),
-                    1); // We only want one address to be returned.
-
-        } catch (IOException e) {
-            // Catch network or other IO problems
-            callback.onAddressFailedResult(mContext.getString(R.string.deviceLocationUtil_geocoder_not_available));
-            return;
-        } catch (IllegalArgumentException e) {
-            // Catch invalid latitude or longitude values
-            callback.onAddressFailedResult(mContext.getString(R.string.deviceLocationUtil_geocoder_invalid_latLong));
-            return;
-        }
-
-        // Handle case where no address is found
-        if (addressList == null || addressList.size() == 0) {
-            callback.onAddressFailedResult(mContext.getString(R.string.deviceLocationUtil_geocoder_address_not_found));
-        } else {
-            // Return the address list
-            callback.onAddressSuccessfulResult(addressList);
-        }
-
-    }
-
-
-    /**
-     * Returns a String containing the requested address element or null if not found
-     *
-     * @param   elementCode A package-defined int constant representing the specific
-     *                      address element to return.
-     * @param   location    A Location object containing a latitude and longitude.
-     *
-     * @return  String containing the requested address element if found, a reason for
-     *          failure if necessary or null if address element doesn't exist.
-     */
-    public String getAddressElement(int elementCode, Location location){
-
-        mGeocoder = new Geocoder(mContext, Locale.getDefault());
-        List<Address> addressList;
-        Address address;
-        String elementString = null;
-
-        try {
-
-            addressList = mGeocoder.getFromLocation(
-                    location.getLatitude(), location.getLongitude(),
-                    1); // We only want one address to be returned.
-
-        } catch (IOException e) {
-            // Catch network or other IO problems
-            return mContext.getString(R.string.deviceLocationUtil_geocoder_not_available);
-        } catch (IllegalArgumentException e) {
-            // Catch invalid latitude or longitude values
-            return mContext.getString(R.string.deviceLocationUtil_geocoder_invalid_latLong);
-        }
-
-        // Handle case where no address is found
-        if (addressList == null || addressList.size() == 0){
-            return mContext.getString(R.string.deviceLocationUtil_geocoder_address_not_found);
-        } else {
-            // Create the Address object from the address list
-            address = addressList.get(0);
-        }
-
-        // Get the specific address element requested by the caller
-        switch (elementCode){
-
-            case AddressCodes.ADMIN_AREA:
-                elementString = address.getAdminArea();
-                break;
-            case AddressCodes.CITY_NAME:
-                elementString = address.getLocality();
-                break;
-            case AddressCodes.COUNTRY_CODE:
-                elementString = address.getCountryCode();
-                break;
-            case AddressCodes.COUNTRY_NAME:
-                elementString = address.getCountryName();
-                break;
-            case AddressCodes.FEATURE_NAME:
-                elementString = address.getFeatureName();
-                break;
-            case AddressCodes.FULL_ADDRESS:
-                elementString = address.toString();
-                break;
-            case AddressCodes.PHONE_NUMBER:
-                elementString = address.getPhone();
-                break;
-            case AddressCodes.POST_CODE:
-                elementString = address.getPostalCode();
-                break;
-            case AddressCodes.PREMISES:
-                elementString = address.getPremises();
-                break;
-            case AddressCodes.STREET_ADDRESS:
-                elementString = address.getThoroughfare();
-                break;
-            case AddressCodes.SUB_ADMIN_AREA:
-                elementString = address.getSubAdminArea();
-                break;
-            case AddressCodes.SUB_THOROUGHFARE:
-                elementString = address.getSubThoroughfare();
-                break;
-            default:
-                elementString = mContext.getString(R.string.deviceLocationUtil_geocoder_invalid_element);
-                break;
-        }
-
-        return elementString;
-    }
-
-
-}// End DeviceLocationUtility class
+}// End EasyLocationUtility class
