+++ DmgLocationUtils/src/main/res/values/strings.xml
