+++ app/build.gradle
@@ -26,5 +26,5 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    compile project(path: ':devicelocationutility')
+    compile project(path: ':DmgLocationUtils')
 }
