+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':devicelocationutility'
+include ':app', ':DmgLocationUtils'
